fix: wg.Done for all goroutines expected but not created
whenever at least 1 of the n expected activities failed to scrape,
the pipeline stage of genImg would block because it created a
wg of size n.

To remediate this, a counter of the number of created go routines
is kept by the genImg pipeline stage. Once the input channel is
closed (no more graphs to turn into images), the stage wg.Done's
n minus the number of created goroutines.